// lib/graphql/generated/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /** An ISO 8601-encoded datetime */
  ISO8601DateTime: { input: any; output: any; }
  Upload: { input: any; output: any; }
  /** A valid URL, transported as a string */
  Url: { input: any; output: any; }
};
// ...
export type Comment = {
  __typename?: 'Comment';
  contents?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  currentUserHasLike?: Maybe<Scalars['Boolean']['output']>;
  discardedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  id: Scalars['ID']['output'];
  isMine?: Maybe<Scalars['Boolean']['output']>;
  isSpoiler: Scalars['Boolean']['output'];
  likeNum?: Maybe<Scalars['Int']['output']>;
  replies?: Maybe<ReplyConnection>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  user?: Maybe<User>;
  userId: Scalars['Int']['output'];
  video?: Maybe<OriginalVideo>;
  videoId: Scalars['Int']['output'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  adminComments: CommentConnection;
  adminCoupons: CouponConnection;
  adminImages: AdminImageConnection;
  adminOpsConfig: OpsConfig;
  adminOriginalVideos: OriginalVideoConnection;
  adminScreenPopup?: Maybe<ScreenPopup>;
  adminScreenPopups: ScreenPopupConnection;
  adminTestMovies: Array<TestMovie>;
  adminUsers: UserConnection;
  adminWithdrawalVideos: Array<WithdrawalVideo>;
  allOriginalVideos: Array<OriginalVideo>;
  category: Category;
  commentReplies: CommentRepliesConnection;
  coupon?: Maybe<Coupon>;
  currentUser?: Maybe<User>;
  diveruOriginalVideos: OriginalVideoConnection;
  firstRecommendationRow: RecommendationRow;
  homeScreens: Array<HomeScreen>;
  jiqooCategory: Category;
  jiqooCommentReplies: CommentRepliesConnection;
  jiqooHomeScreens: Array<HomeScreen>;
  jiqooOriginalVideo?: Maybe<OriginalVideo>;
  jiqooOriginalVideos: OriginalVideoConnection;
  jiqooRootCategories: Array<RootCategory>;
  jiqooScreenPopups: ScreenPopupConnection;
  jiqooTrailerVideo?: Maybe<TrailerVideo>;
  jiqooTrailerVideos: TrailerVideoConnection;
  jiqooVideoComments: VideoCommentsConnection;
  jiqooVideoRecommends?: Maybe<VideoRecommendsConnection>;
  notifications: UserNotificationConnection;
  originalVideo?: Maybe<OriginalVideo>;
  originalVideos: OriginalVideoConnection;
  payment: Payment;
  playlist: Playlist;
  playlists: Array<Playlist>;
  promotionCode?: Maybe<PromotionCode>;
  recommendationRow: RecommendationRow;
  recommendationRowList: RecommendationRowList;
  recommendedSearchKeywords: Array<RecommendedSearchKeyword>;
  rootCategories: Array<RootCategory>;
  screenPopups: ScreenPopupConnection;
  trailerVideo?: Maybe<TrailerVideo>;
  trailerVideos: TrailerVideoConnection;
  trialVideos: Array<OriginalVideo>;
  tvSearchCategories: Array<TvSearchCategory>;
  userNotificationSettings: Array<UserNotificationSetting>;
  videoAiRecommends?: Maybe<VideoRecommendsConnection>;
  videoComments: VideoCommentsConnection;
  videoHistories: VideoHistoryConnection;
  videoRatings: VideoRatingConnection;
  videoRecommends?: Maybe<VideoRecommendsConnection>;
  withdrawalOriginalVideos: Array<OriginalVideo>;
  withdrawalVideos: Array<TrailerVideo>;
};
// ...
export type User = {
  __typename?: 'User';
  age?: Maybe<Scalars['Int']['output']>;
  appsflyerId?: Maybe<Scalars['String']['output']>;
  autoplayVideos: Scalars['Boolean']['output'];
  avatar?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  dateOfBirth?: Maybe<Scalars['ISO8601Date']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  freeTrialEndDate?: Maybe<Scalars['ISO8601Date']['output']>;
  gender?: Maybe<GenderEnum>;
  hasCancelledSubscription?: Maybe<Scalars['Boolean']['output']>;
  hasLineLinked: Scalars['Boolean']['output'];
  hasSubscription?: Maybe<Scalars['Boolean']['output']>;
  histories?: Maybe<OriginalVideoConnection>;
  id: Scalars['ID']['output'];
  isSelf?: Maybe<Scalars['Boolean']['output']>;
  language?: Maybe<Scalars['String']['output']>;
  likes?: Maybe<LikeConnection>;
  mailDesired?: Maybe<Scalars['Boolean']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  payment: Payment;
  points?: Maybe<Scalars['Int']['output']>;
  provider?: Maybe<Scalars['String']['output']>;
  qualitySetting?: Maybe<Scalars['Int']['output']>;
  speedSetting?: Maybe<Scalars['String']['output']>;
  stripeCustomerId?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  userStatus?: Maybe<Scalars['String']['output']>;
  withdrawal?: Maybe<Withdrawal>;
};
// ...
export type VideoCommentsConnection = {
  __typename?: 'VideoCommentsConnection';
  allCount?: Maybe<Scalars['Int']['output']>;
  /** A list of edges. */
  edges?: Maybe<Array<VideoCommentsEdge>>;
  id: Scalars['ID']['output'];
  /** A list of nodes. */
  nodes?: Maybe<Array<Comment>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};

/** An edge in a connection. */
export type VideoCommentsEdge = {
  __typename?: 'VideoCommentsEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Comment>;
};
// ...
export type GetVideoCommentsQueryVariables = Exact<{
  id: Scalars['ID']['input'];
  first?: InputMaybe<Scalars['Int']['input']>;
  after?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetVideoCommentsQuery = { __typename?: 'Query', videoComments: { __typename?: 'VideoCommentsConnection', id: string, allCount?: number | null, pageInfo?: { __typename?: 'PageInfo', startCursor?: string | null, endCursor?: string | null, hasNextPage: boolean, hasPreviousPage: boolean } | null, edges?: Array<{ __typename?: 'VideoCommentsEdge', cursor: string, node?: { __typename?: 'Comment', id: string, contents?: string | null, createdAt?: any | null, likeNum?: number | null, user?: { __typename?: 'User', id: string, name?: string | null, avatar?: string | null } | null } | null }> | null } };
// ...
export const GetVideoCommentsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"getVideoComments"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"first"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"after"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"videoComments"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}},{"kind":"Argument","name":{"kind":"Name","value":"first"},"value":{"kind":"Variable","name":{"kind":"Name","value":"first"}}},{"kind":"Argument","name":{"kind":"Name","value":"after"},"value":{"kind":"Variable","name":{"kind":"Name","value":"after"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"allCount"}},{"kind":"Field","name":{"kind":"Name","value":"pageInfo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"startCursor"}},{"kind":"Field","name":{"kind":"Name","value":"endCursor"}},{"kind":"Field","name":{"kind":"Name","value":"hasNextPage"}},{"kind":"Field","name":{"kind":"Name","value":"hasPreviousPage"}}]}},{"kind":"Field","name":{"kind":"Name","value":"edges"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"cursor"}},{"kind":"Field","name":{"kind":"Name","value":"node"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"contents"}},{"kind":"Field","name":{"kind":"Name","value":"user"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"avatar"}}]}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"likeNum"}}]}}]}}]}}]}}]} as unknown as DocumentNode<GetVideoCommentsQuery, GetVideoCommentsQueryVariables>;